+++ android/tools/render-guide.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # vim: set et ts=4 sw=4:
 # =============================================================================
-# Libellus Potionis "Potillus" -- Privacy-Friendly Alcohol Tracker
+# Libellus Potionis -- Privacy-Friendly Alcohol Tracker
 # Copyright (c) 2026 Martin A. Godisch <android@godisch.de>
 # =============================================================================
 #